formatting dash app w/ best practices
